+++ m1/asr/__init__.py
@@ -1,1 +1,1 @@
-"""ASR helpers."""
+"""ASR helpers."""\nfrom .service import ASRService, TranscriptSegment\n\n__all__ = ["ASRService", "TranscriptSegment"]